+++ exercises/openmp/c/tasks/pngwriter.h
@@ -1,6 +1,6 @@
 #ifndef PNGWRITER_H_
 #define PNGWRITER_H_
 
-int save_png(long *data, const int nx, const int ny, const char *fname);
+int save_png(long long *data, const long long nx, const long long ny, const char *fname);
 
 #endif
